pyproject.toml
[tool.poetry]
name = "rjwedding-backend"
version = "0.1.0"
description = ""
authors = ["burtonrj <burtonrossj@gmail.com>"]
readme = "README.md"
packages = [{include = "rjwedding_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
python-decouple = "^3.8"
pandas = "^2.0.1"
hypercorn = "^0.14.3"
odmantic = "^0.9.2"
boto3 = "^1.26.133"
docker = "^6.1.2"
python-multipart = "^0.0.6"
httpx = "^0.24.0"
pytest-asyncio = "^0.21.0"


[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
mypy = "*"
pre-commit = "*"
black = "*"
flake8 = "*"
isort = "*"


[tool.pytest.ini_options]
asyncio_mode = "auto"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
